moved header to a separate component
