+++ src/views/ModelDetails.vue
@@ -1,41 +1,6 @@
 <template>
-  <el-header class="hidden-md-and-down">
-    <el-row :gutter="10">
-      <el-col :xs="4" :span="2" :xl="1">
-        <router-link to="/">
-          <el-button type="primary" plain>
-            <font-awesome-icon :icon="['fas', 'arrow-left']" />
-          </el-button>
-        </router-link>
-      </el-col>
-      <el-col class="hidden-sm-and-down" :span="2">
-        <h1 class="logo">
-          <span>ID</span>
-          <span>AR</span>
-        </h1>
-      </el-col>
-      <el-col
-        :span="22"
-        :xs="{ span: 19, offset: 1 }"
-        :md="{ span: 12, offset: 8 }"
-        :xl="{ span: 8, offset: 13 }"
-      >
-        <el-input
-          v-model="searchValue"
-          placeholder="Search for model"
-          clearable
-        >
-          <template #prefix>
-            <font-awesome-icon
-              style="margin-left: 5px"
-              :icon="['fas', 'search']"
-            />
-          </template>
-        </el-input>
-      </el-col>
-    </el-row>
-  </el-header>
   <el-container>
+    <Header />
     <el-main v-if="modelMetaData">
       <el-image fit="contain" :src="modelMetaData.modelImagePath" />
       <el-row align="middle">
@@ -59,7 +24,12 @@
         </el-col>
       </el-row>
     </el-main>
-    <p class="not-found" v-else>Requested model was not found</p>
+    <div class="not-found" v-else>
+      <p>Requested model was not found</p>
+      <router-link to="/">
+        <el-button>Go back to main page</el-button>
+      </router-link>
+    </div>
     <Footer />
   </el-container>
 </template>
@@ -69,13 +39,13 @@ import { computed, defineComponent, inject, ref } from "vue";
 import { useRoute } from "vue-router";
 import { ModelsRefInjectKey } from "@/symbols";
 
+import Header from "@/components/Header.vue";
 import Footer from "@/components/Footer.vue";
-
 import { useXR } from "@/composables/webxr/composables/useXR";
 
 export default defineComponent({
   name: "Model Details",
-  components: { Footer },
+  components: { Header, Footer },
   setup() {
     const route = useRoute();
     const models = inject(ModelsRefInjectKey);
@@ -88,26 +58,20 @@ export default defineComponent({
     const toastMessage = ref("");
     const { isXrSupported, startAR } = useXR(toastMessage);
 
-    const searchValue = ref("");
-
     return {
       modelMetaData,
       startAR,
       isXrSupported,
-      searchValue,
     };
   },
 });
 </script>
 
 <style lang="scss" scoped>
-.el-container {
-  background-color: rgb(245, 245, 245);
-  height: calc(100vh - 32px);
-}
-
 .el-main {
-  --el-main-padding: 0;
+  background-color: rgb(245, 245, 245);
+  min-height: 100vh;
+  --el-main-padding: 60px 0 0;
   @media only screen and (min-width: 1200px) {
     --el-main-padding: 70px 200px 32px;
   }
@@ -136,41 +100,13 @@ export default defineComponent({
   }
 }
 
-.el-header {
-  --el-header-padding: 10px 10px;
-  @media only screen and (min-width: 1200px) {
-    --el-header-padding: 10px 200px;
-  }
-  top: 0px;
-  left: 0px;
-  position: fixed;
-  width: 100%;
-  background-color: white;
-  z-index: 100;
-
-  .el-button {
-    padding: 12px 13px;
-  }
-
-  .logo {
-    margin: -5px 0 0;
-    line-height: 1;
-    font-size: 1.1rem;
-    span:first-child {
-      font-size: 3em;
-    }
-    span:last-child {
-      font-size: 0.75em;
-      position: relative;
-      left: -0.3rem;
-    }
-  }
-}
-
 .not-found {
+  --not-found-padding-top: 100px;
   text-align: center;
   width: 100%;
-  margin-top: 100px;
+  height: calc(100vh - var(--not-found-padding-top));
+  padding-top: var(--not-found-padding-top);
   font-size: 2rem;
+  background-color: rgb(245, 245, 245);
 }
 </style>
